Fixes stuff that broke building with llvm 3.9
